// apps/jobs/web/src/lib/api/generated/model/companyInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyInputSegment } from './companyInputSegment';
import type { CompanyInputSizeRange } from './companyInputSizeRange';
import type { CompanyInputHq } from './companyInputHq';

export interface CompanyInput {
  /**
   * @minLength 2
   * @maxLength 80
   */
  name: string;
  /** @maxLength 500 */
  logoUrl?: string;
  /** @maxLength 500 */
  coverUrl?: string;
  /** @maxLength 500 */
  website?: string;
  /** @maxLength 4000 */
  about?: string;
  segment?: CompanyInputSegment;
  sizeRange?: CompanyInputSizeRange;
  hq?: CompanyInputHq;
  /** @maxLength 40 */
  ownerHandle?: string;
}